+++ muxers/mplex/tests/async_write.rs
@@ -22,7 +22,6 @@ use futures::future::poll_fn;
 use futures::{channel::oneshot, prelude::*};
 use libp2p_core::{upgrade, StreamMuxer, Transport};
 use libp2p_tcp::TcpTransport;
-use std::sync::Arc;
 
 #[test]
 fn async_write() {
@@ -60,10 +59,7 @@ fn async_write() {
             .await
             .unwrap();
 
-        let mut outbound_token = client.open_outbound();
-        let mut outbound = poll_fn(|cx| client.poll_outbound(cx, &mut outbound_token))
-            .await
-            .unwrap();
+        let mut outbound = poll_fn(|cx| client.poll_outbound(cx)).await.unwrap();
 
         let mut buf = Vec::new();
         outbound.read_to_end(&mut buf).await.unwrap();
@@ -75,16 +71,8 @@ fn async_write() {
         let mut transport = TcpTransport::default()
             .and_then(move |c, e| upgrade::apply(c, mplex, e, upgrade::Version::V1));
 
-        let client = Arc::new(transport.dial(rx.await.unwrap()).unwrap().await.unwrap());
-        let mut inbound = loop {
-            if let Some(s) = poll_fn(|cx| client.poll_event(cx))
-                .await
-                .unwrap()
-                .into_inbound_substream()
-            {
-                break s;
-            }
-        };
+        let client = transport.dial(rx.await.unwrap()).unwrap().await.unwrap();
+        let mut inbound = poll_fn(|cx| client.poll_inbound(cx)).await.unwrap();
         inbound.write_all(b"hello world").await.unwrap();
 
         // The test consists in making sure that this flushes the substream.