+++ protocols/relay/src/v2.rs
@@ -34,7 +34,8 @@ pub use protocol::{
     inbound_hop::FatalUpgradeError as InboundHopFatalUpgradeError,
     inbound_stop::FatalUpgradeError as InboundStopFatalUpgradeError,
     outbound_hop::FatalUpgradeError as OutboundHopFatalUpgradeError,
-    outbound_stop::FatalUpgradeError as OutboundStopFatalUpgradeError,
+    outbound_stop::FatalUpgradeError as OutboundStopFatalUpgradeError, HOP_PROTOCOL_NAME,
+    STOP_PROTOCOL_NAME,
 };
 
 /// The ID of an outgoing / incoming, relay / destination request.