+++ muxers/mplex/tests/two_peers.rs
@@ -22,7 +22,6 @@ use futures::{channel::oneshot, prelude::*};
 use libp2p_core::muxing::StreamMuxerExt;
 use libp2p_core::{upgrade, Transport};
 use libp2p_tcp::TcpTransport;
-use std::sync::Arc;
 
 #[test]
 fn client_to_server_outbound() {
@@ -73,7 +72,7 @@ fn client_to_server_outbound() {
             .and_then(move |c, e| upgrade::apply(c, mplex, e, upgrade::Version::V1))
             .boxed();
 
-        let mut client = Arc::new(transport.dial(rx.await.unwrap()).unwrap().await.unwrap());
+        let mut client = transport.dial(rx.await.unwrap()).unwrap().await.unwrap();
         let mut inbound = client.next_inbound().await.unwrap();
         inbound.write_all(b"hello world").await.unwrap();
         inbound.close().await.unwrap();
@@ -108,17 +107,15 @@ fn client_to_server_inbound() {
 
         tx.send(addr).unwrap();
 
-        let mut client = Arc::new(
-            transport
-                .next()
-                .await
-                .expect("some event")
-                .into_incoming()
-                .unwrap()
-                .0
-                .await
-                .unwrap(),
-        );
+        let mut client = transport
+            .next()
+            .await
+            .expect("some event")
+            .into_incoming()
+            .unwrap()
+            .0
+            .await
+            .unwrap();
 
         let mut inbound = client.next_inbound().await.unwrap();