+++ protocols/gossipsub/CHANGELOG.md
@@ -1,4 +1,4 @@
-# 0.40.0 [unreleased]
+# 0.40.0
 
 - Update to `libp2p-swarm` `v0.38.0`.